requirements.txt
aiogram==3.4.1
httpx
python-dotenv
aiosqlite
openai
apscheduler
trafilatura
